Change TabCard to Tabs, remove TabCard, move renderTitles
